pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "groww-pulse"
version = "0.1.0"
description = "Weekly Product Review Pulse — automated review analysis and delivery"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "pydantic[email]>=2.0",
    "pyyaml>=6.0",
    "structlog>=24.0",
    "tenacity>=8.0",
    "mcp>=1.0",
    "httpx>=0.27",
    "google-play-scraper>=1.2",
    "langdetect>=1.0",
    "python-dateutil>=2.8",
    "openai>=1.0",
    "groq>=0.9",
    "fastmcp>=2.0",
    "python-dotenv>=1.0",
    "sentence-transformers>=3.0",
    "google-api-python-client>=2.0",
    "google-auth-httplib2>=0.2",
    "google-auth-oauthlib>=1.0",
    "jinja2>=3.1",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "numpy>=1.24",
    "umap-learn>=0.5",
    "hdbscan>=0.8",
    "scikit-learn>=1.4",
]

[project.scripts]
pulse = "pulse.phase_0.cli.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "mypy>=1.10",
    "ruff>=0.5",
    "pytest-freezer>=0.4",
]

[tool.hatch.build.targets.wheel]
packages = ["src/pulse"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "N", "B"]
ignore = ["B008"]  # typer uses function calls in defaults

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = "src"
explicit_package_bases = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
